feat: Add GitHub Projects v2 automation
- Add project-sync.yml workflow for automatic issue/PR sync
- Add project-backfill.yml for manual bulk sync
- Add comprehensive documentation in docs/
- Add bootstrap and field mapping guides in scripts/
- Update issue templates with project fields
- Update PR template with project sync requirements
- Add CODEOWNERS protection for workflows

Features:
- Automatic status updates based on events
- Label-to-field mapping (type, priority, sprint)
- Estimate parsing from issue bodies
- CI integration with status updates
- Dry-run support for testing
- Error handling with retry logic
- Audit trails and security guardrails
- Reusable across repositories
