+++ .github/ISSUE_TEMPLATE/bug.md
@@ -1,0 +1,46 @@
+---
+name: Bug Report
+about: Report a bug or issue
+title: "[BUG] "
+labels: ["type:bug"]
+---
+
+## Description
+Brief description of the bug
+
+## Steps to Reproduce
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
+
+## Expected Behavior
+What should happen
+
+## Actual Behavior
+What actually happens
+
+## Severity
+- [ ] P0 (Critical - blocks work, no workaround)
+- [ ] P1 (High - major issue, workaround exists)
+- [ ] P2 (Medium - minor issue, cosmetic)
+- [ ] P3 (Low - very minor, edge case)
+
+## Environment
+- **Browser**: [e.g., Chrome 91, Firefox 89]
+- **OS**: [e.g., Windows 10, macOS 11]
+- **Device**: [e.g., Desktop, Mobile]
+- **Version**: [e.g., v1.2.3]
+
+## Screenshots
+If applicable, add screenshots to help explain the problem
+
+## Additional Context
+Add any other context about the problem here, such as:
+- Error messages
+- Console logs
+- Related issues
+- Workarounds tried
+
+## Estimate
+Estimate:
