+++ .github/ISSUE_TEMPLATE/feature.md
@@ -1,0 +1,35 @@
+---
+name: Feature Request
+about: Request a new feature
+title: "[FEATURE] "
+labels: ["type:feature"]
+---
+
+## Description
+Brief description of the feature request
+
+## User Story
+As a [user type], I want [goal] so that [benefit]
+
+## Acceptance Criteria
+- [ ] Criteria 1
+- [ ] Criteria 2
+- [ ] Criteria 3
+
+## Priority
+- [ ] P0 (Critical - blocks other work)
+- [ ] P1 (High - important for next release)
+- [ ] P2 (Medium - should have eventually)
+- [ ] P3 (Low - nice to have)
+
+## Estimate
+Estimate:
+
+## Sprint
+Sprint:
+
+## Additional Context
+Add any other context, screenshots, or examples about the feature request here.
+
+## Related Issues
+Link to any related issues or discussions
