+++ app/api/admin/seed-database/route.ts
@@ -1,0 +1,18 @@
+import { NextResponse } from 'next/server';
+import { execSync } from 'child_process';
+
+export async function POST() {
+  try {
+    execSync('npx prisma db seed', {
+      stdio: 'inherit',
+      cwd: process.cwd()
+    });
+
+    return NextResponse.json({ success: true });
+  } catch (error) {
+    return NextResponse.json({
+      success: false,
+      error: error instanceof Error ? error.message : 'Seeding failed'
+    }, { status: 500 });
+  }
+}
